Realize Strategy should come with type metadata
Specifically, it should say if the type is an adt, and if the type is a
tmf, it should say the pre-mapped tmf type. Without the pre-mapped tmf
type, it is not possible to operate on the tmf in the same way as if we
had been in the original language, because the wrapping tmf is not
guaranteed to, and usually won't, have corresponding behavior to the
wrapped one.