WIP, probably revert me
Realizing that in order for generated code to name its own sort ids, it
needs to have dependencies on a bunch of crates that it shouldn't have
dependencies on. Instead, maybe it should name its own index in the list
of all sorts for the langspec of the current heap.