WIP; realize words::Implements needed for TMFs
It can be implemented for the skeletons. The key insight is that some of
the TMFs in the extended language correspond to TMFs in the original
language, whereas others are parts of mapped ADTs. In the context of
bridging between languages, it is right and proper that these TMFs
should be treated differently.