Remove confusion of responsibilities between...
bridge, term_specialized_impl. Part of term_specialized_impl is just a
bridge from a language to itself -- that is, a special case of bridge.

This does not reduce the amount of code significantly because the two
were already calling the same functions and/or they were doing trivial
things. However, it arguably makes the situation less confusing.